requirements.txt
beautifulsoup4==4.12.3
Django==5.0.3
django-allauth==0.61.1
djangorestframework==3.14.0
uvicorn==0.27.1
whitenoise==6.6.0
python-decouple==3.8